Game Base
Has mobs in this case, stars that shoot across the screen and a ground.
The next step should be to add the hero. I’m thinking I can use a
square or something until the artwork is done. My next step will be to
add controls so the square can duck and jump.