Adding OpenGl Project bones
It’s super basic I’m building out the Triangle class and following
Google’s OpenGL example. My hope is to turn the triangle into a star
and then use physics to create and shoot the stars across the screen. I
have made the background blue-ish so it can stand in as a sky. I also
have plans to add the ground.